remove focuszone code
