+++ vnext/Microsoft.ReactNative/Views/SliderViewManager.cpp
@@ -27,7 +27,6 @@ class SliderShadowNode : public ShadowNodeBase {
   SliderShadowNode() = default;
   void createView() override;
   void updateProperties(winrt::Microsoft::ReactNative::JSValueObject &props) override;
-  bool NeedsForceLayout();
 };
 
 void SliderShadowNode::createView() {
@@ -38,10 +37,6 @@ void SliderShadowNode::updateProperties(winrt::Microsoft::ReactNative::JSValueOb
   m_updating = true;
   Super::updateProperties(props);
   m_updating = false;
-}
-
-bool SliderShadowNode::NeedsForceLayout() {
-  return true;
 }
 
 SliderViewManager::SliderViewManager(const Mso::React::IReactContext &context) : Super(context) {}
@@ -89,10 +84,6 @@ bool SliderViewManager::UpdateProperty(
     return Super::UpdateProperty(nodeToUpdate, propertyName, propertyValue);
   }
   return true;
-}
-
-YGMeasureFunc SliderViewManager::GetYogaCustomMeasureFunc() const {
-  return DefaultYogaSelfMeasureFunc;
 }
 
 } // namespace Microsoft::ReactNative
